+++ CHANGES.markdown
@@ -1,3 +1,7 @@
+## Changes in next
+ - `Functor` instances for `Handler` and `ArrowMonad`. `Applicative`,
+   `Alternative`, and `MonadPlus` instances for `ArrowMonad`.
+
 ## Changes in 0.4.0
  - Removed all `Generic` and `Generic1` instances. These have been moved to the
    `generic-deriving` library.